Extract HomeScreen to a dedicated file
