+++ ScopedStorage/app/src/main/java/com/samples/storage/scopedstorage/MainActivity.kt
@@ -4,25 +4,15 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.Divider
 import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.Icon
-import androidx.compose.material.ListItem
-import androidx.compose.material.Scaffold
 import androidx.compose.material.Text
-import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.navigation.NavController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
@@ -42,9 +32,7 @@ class MainActivity : ComponentActivity() {
                     navController = navController,
                     startDestination = HomeRoute
                 ) {
-                    composable(HomeRoute) {
-                        HomeScreen(navController)
-                    }
+                    composable(HomeRoute) { HomeScreen(navController) }
 
                     composable(Demos.AddMediaFile.route) { AddMediaFileScreen() }
                     composable(Demos.CaptureMediaFile.route) { NotAvailableYetScreen() }
@@ -60,31 +48,6 @@ class MainActivity : ComponentActivity() {
             }
         }
     }
-}
-
-const val HomeRoute = "home"
-
-@ExperimentalMaterialApi
-@Composable
-fun HomeScreen(navController: NavController) {
-    Scaffold(
-        topBar = {
-            TopAppBar(title = { Text(stringResource(R.string.app_name)) })
-        },
-        content = { paddingValues ->
-            LazyColumn(Modifier.padding(paddingValues)) {
-                items(Demos.list) { demo ->
-                    ListItem(
-                        modifier = Modifier.clickable { navController.navigate(demo.route) },
-                        text = { Text(stringResource(demo.name)) },
-                        secondaryText = { Text(stringResource(demo.description)) },
-                        icon = { Icon(demo.icon, contentDescription = null) }
-                    )
-                    Divider()
-                }
-            }
-        }
-    )
 }
 
 @Composable
